restructuring, adding tests, updating formatting function
